"Added GestureDetector to remove keyboard on tap and initialized ConnectivityController in Scaffold builder"
This commit adds a GestureDetector widget to the Scaffold's body, which removes the keyboard when the user taps outside of any text fields. It also initializes the ConnectivityController in the Scaffold's builder. The previous implementation initialized the ConnectivityController in the builder of the Builder widget, which is now wrapped in the GestureDetector widget.

Signed-off-by: Your Name <kareem01229526319@gmail.com>